requirements.txt
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
streamlit>=1.28.0
plotly>=5.15.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
redis>=4.6.0
pymongo>=4.4.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
nltk>=3.8.1
jieba>=0.42.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
gym>=0.26.0
stable-baselines3>=2.1.0
tensorboard>=2.13.0
